Move all screens to ui package
