+++ mobile-ui/src/main/AndroidManifest.xml
@@ -22,7 +22,7 @@
                 android:value="@string/google_maps_api_key"/>
 
         <activity
-                android:name=".splash.SplashActivity"
+                android:name="com.alexzh.imbarista.ui.splash.SplashActivity"
                 android:theme="@style/AppTheme.SplashScreen">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN"/>
@@ -31,20 +31,20 @@
             </intent-filter>
         </activity>
         <activity
-                android:name=".login.LoginActivity"
+                android:name="com.alexzh.imbarista.ui.login.LoginActivity"
                 android:label="@string/activity_login_title"
                 android:windowSoftInputMode="stateVisible|adjustResize"/>
         <activity
-                android:name=".createaccount.CreateAccountActivity"
+                android:name="com.alexzh.imbarista.ui.createaccount.CreateAccountActivity"
                 android:label="@string/activity_create_account_title"
                 android:windowSoftInputMode="stateVisible|adjustResize">
         </activity>
-        <activity android:name=".home.HomeActivity">
+        <activity android:name="com.alexzh.imbarista.ui.home.HomeActivity">
         </activity>
-        <activity android:name=".coffeedrinkdetails.CoffeeDetailsActivity">
+        <activity android:name="com.alexzh.imbarista.ui.coffeedrinkdetails.CoffeeDetailsActivity">
         </activity>
         <activity
-                android:name=".settings.SettingsActivity"
+                android:name="com.alexzh.imbarista.ui.settings.SettingsActivity"
                 android:label="@string/activity_settings_title">
         </activity>
     </application>
