+++ mobile-ui/src/main/java/com/alexzh/imbarista/ui/coffeedrinks/adapter/CoffeeDrinksAdapter.kt
@@ -1,4 +1,4 @@
-package com.alexzh.imbarista.coffeedrinks.adapter
+package com.alexzh.imbarista.ui.coffeedrinks.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
