+++ mobile-ui/src/main/java/com/alexzh/imbarista/ui/home/HomeActivity.kt
@@ -1,4 +1,4 @@
-package com.alexzh.imbarista.home
+package com.alexzh.imbarista.ui.home
 
 import android.content.Context
 import android.content.Intent
@@ -7,10 +7,10 @@ import android.os.Bundle
 import android.view.MenuItem
 import androidx.fragment.app.Fragment
 import com.alexzh.imbarista.R
-import com.alexzh.imbarista.coffeedrinks.CoffeeDrinksFragment
-import com.alexzh.imbarista.map.TomTomMapFragment
-import com.alexzh.imbarista.profile.ProfileFragment
-import com.alexzh.imbarista.settings.SettingsActivity
+import com.alexzh.imbarista.ui.coffeedrinks.CoffeeDrinksFragment
+import com.alexzh.imbarista.ui.map.TomTomMapFragment
+import com.alexzh.imbarista.ui.profile.ProfileFragment
+import com.alexzh.imbarista.ui.settings.SettingsActivity
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import kotlinx.android.synthetic.main.activity_home.*
