+++ mobile-ui/src/main/res/layout/activity_create_account.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".createaccount.CreateAccountActivity">
+        tools:context=".ui.createaccount.CreateAccountActivity">
 
     <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
