+++ mobile-ui/src/main/res/layout/activity_home.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".coffeedrinks.CoffeeDrinksFragment">
+        tools:context=".ui.coffeedrinks.CoffeeDrinksFragment">
 
     <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
