+++ mobile-ui/src/main/res/layout/activity_login.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".login.LoginActivity">
+        tools:context=".ui.login.LoginActivity">
 
     <androidx.appcompat.widget.Toolbar
             android:id="@+id/toolbar"
