+++ mobile-ui/src/main/res/layout/fragment_coffee_drinks.xml
@@ -5,7 +5,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".coffeedrinks.CoffeeDrinksFragment">
+        tools:context=".ui.coffeedrinks.CoffeeDrinksFragment">
 
     <androidx.recyclerview.widget.RecyclerView
             android:id="@+id/recyclerView"
